+++ packages/@expressive-code/core/src/common/annotation.ts
@@ -94,6 +94,12 @@ export type InlineStyleAnnotationOptions = AnnotationBaseOptions & {
 	 */
 	color?: string | undefined
 	/**
+	 * The background color of the annotation. This is expected to be a hex color string,
+	 * e.g. `#888`. Using CSS variables or other color formats is possible, but prevents
+	 * automatic color contrast checks from working.
+	 */
+	bgColor?: string | undefined
+	/**
 	 * Whether the annotation should be rendered in italics.
 	 */
 	italic?: boolean | undefined
@@ -105,6 +111,10 @@ export type InlineStyleAnnotationOptions = AnnotationBaseOptions & {
 	 * Whether the annotation should be rendered with an underline.
 	 */
 	underline?: boolean | undefined
+	/**
+	 * Whether the annotation should be rendered with a strikethrough.
+	 */
+	strikethrough?: boolean | undefined
 	/**
 	 * Inline styles can be theme-dependent, which allows plugins like syntax highlighters to
 	 * style the same code differently depending on the theme.
@@ -139,18 +149,22 @@ export type InlineStyleAnnotationOptions = AnnotationBaseOptions & {
 export class InlineStyleAnnotation extends ExpressiveCodeAnnotation {
 	name: string
 	color: string | undefined
+	bgColor: string | undefined
 	italic: boolean
 	bold: boolean
 	underline: boolean
+	strikethrough: boolean
 	styleVariantIndex: number | undefined
 
-	constructor({ color, italic = false, bold = false, underline = false, styleVariantIndex, ...baseOptions }: InlineStyleAnnotationOptions) {
+	constructor({ color, bgColor, italic = false, bold = false, underline = false, strikethrough = false, styleVariantIndex, ...baseOptions }: InlineStyleAnnotationOptions) {
 		super(baseOptions)
 		this.name = 'Inline style'
 		this.color = color
+		this.bgColor = bgColor
 		this.italic = italic
 		this.bold = bold
 		this.underline = underline
+		this.strikethrough = strikethrough
 		this.styleVariantIndex = styleVariantIndex
 	}
 
@@ -160,9 +174,13 @@ export class InlineStyleAnnotation extends ExpressiveCodeAnnotation {
 		const addStylesForVariantIndex = (variantIndex: number) => {
 			const varPrefix = `--${variantIndex}`
 			if (this.color) newStyles.set(varPrefix, this.color)
+			if (this.bgColor) newStyles.set(`${varPrefix}bg`, this.bgColor)
 			if (this.italic) newStyles.set(`${varPrefix}fs`, 'italic')
 			if (this.bold) newStyles.set(`${varPrefix}fw`, 'bold')
-			if (this.underline) newStyles.set(`${varPrefix}td`, 'underline')
+			const textDecorations = []
+			if (this.underline) textDecorations.push('underline')
+			if (this.strikethrough) textDecorations.push('line-through')
+			if (textDecorations.length) newStyles.set(`${varPrefix}td`, textDecorations.join(' '))
 		}
 		const variantIndices = this.styleVariantIndex !== undefined ? [this.styleVariantIndex] : styleVariants.map((_, i) => i)
 		variantIndices.forEach(addStylesForVariantIndex)