+++ packages/@expressive-code/core/src/internal/render-block.ts
@@ -93,6 +93,8 @@ export async function renderBlock({
 	const lines = codeBlock.getLines()
 	const renderedAstLines: Element[] = []
 	const renderEmptyLine = getRenderEmptyLineFn({ gutterElements, ...baseContext })
+	const { wrap = false, preserveIndent = true, hangingIndent = 0 } = codeBlock.props
+
 	type ClassNames = string | string[]
 	const blockClasses: ClassNames[] = []
 	const addClassesToRenderedBlock = (classNames: ClassNames) => blockClasses.push(classNames)
@@ -112,9 +114,11 @@ export async function renderBlock({
 				...baseContext,
 			}),
 		}
-		// Add indent information if wrapping is enabled and preserveIndent has not been disabled
-		if (codeBlock.props.wrap && codeBlock.props.preserveIndent !== false) {
-			const indent = line.text.match(/^\s*/)?.[0].length ?? 0
+		// Add indent information if wrapping is enabled and the configuration
+		// either requests preserving indent or rendering a hanging indent
+		if (wrap && (preserveIndent || hangingIndent > 0)) {
+			const baseIndent = preserveIndent ? line.text.match(/^\s*/)?.[0].length ?? 0 : 0
+			const indent = baseIndent + hangingIndent
 			if (indent > 0) setInlineStyle(lineRenderData.lineAst, '--ecIndent', `${indent}ch`)
 		}
 		// Apply additional classes to the line element (if any)
@@ -182,11 +186,10 @@ function applyDefaultProps(codeBlock: ExpressiveCodeBlock, config: ExpressiveCod
 			Object.assign(mergedDefaults, overridesByLang[key])
 		}
 	})
-	// Apply the merged defaults to the code block
+	// Apply the merged default values to undefined code block props
 	const defaultKeys = Object.keys(mergedDefaults) as (keyof ExpressiveCodeBlock['props'])[]
-	defaultKeys.forEach((key) => {
-		if (codeBlock.props[key] === undefined) codeBlock.props[key] = mergedDefaults[key]
-	})
+	const undefinedValueKeys = defaultKeys.filter((key) => codeBlock.props[key] === undefined)
+	Object.assign(codeBlock.props, Object.fromEntries(undefinedValueKeys.map((key) => [key, mergedDefaults[key]])))
 }
 
 export interface ExpressiveCodeProcessingState {