+++ packages/@expressive-code/plugin-collapsible-sections/src/styles.ts
@@ -1,4 +1,4 @@
-import { PluginStyleSettings, ResolverContext, codeLineClass, setAlpha } from '@expressive-code/core'
+import { PluginStyleSettings, ResolverContext, codeLineClass, createInlineSvgUrl, setAlpha } from '@expressive-code/core'
 
 export const collapsibleSectionClass = 'ec-section'
 
@@ -85,6 +85,20 @@ export interface CollapsibleSectionsStyleSettings {
 	 * ({ theme }) => setAlpha(theme.colors['editor.foldBackground'], 0.1) || 'rgb(84 174 255 / 10%)'
 	 */
 	openBackgroundColorCollapsible: string
+	/**
+	 * An inline SVG URL for the expand icon.
+	 *
+	 * Expects a string in the format `url("data:image/svg+xml,...")`, which can
+	 * be generated from the contents of an SVG file using {@link createInlineSvgUrl}.
+	 */
+	expandIcon: string
+	/**
+	 * An inline SVG URL for the collapse icon.
+	 *
+	 * Expects a string in the format `url("data:image/svg+xml,...")`, which can
+	 * be generated from the contents of an SVG file using {@link createInlineSvgUrl}.
+	 */
+	collapseIcon: string
 }
 
 export const collapsibleSectionsStyleSettings = new PluginStyleSettings({
@@ -105,19 +119,25 @@ export const collapsibleSectionsStyleSettings = new PluginStyleSettings({
 			openBackgroundColor: 'transparent',
 			openBackgroundColorCollapsible: ({ theme }) => setAlpha(theme.colors['editor.foldBackground'], 0.1) || 'rgb(84 174 255 / 10%)',
 			openBorderColor: 'transparent',
+			// Icon source: Octicons (MIT licensed)
+			expandIcon: createInlineSvgUrl([
+				`<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'>`,
+				`<path d='m8.177.677 2.896 2.896a.25.25 0 0 1-.177.427H8.75v1.25a.75.75 0 0 1-1.5 0V4H5.104a.25.25 0 0 1-.177-.427L7.823.677a.25.25 0 0 1 .354 0ZM7.25 10.75a.75.75 0 0 1 1.5 0V12h2.146a.25.25 0 0 1 .177.427l-2.896 2.896a.25.25 0 0 1-.354 0l-2.896-2.896A.25.25 0 0 1 5.104 12H7.25v-1.25Zm-5-2a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM6 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 6 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM12 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 12 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5Z'/>`,
+				`</svg>`,
+			]),
+			// Icon source: Octicons (MIT licensed)
+			collapseIcon: createInlineSvgUrl([
+				`<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'>`,
+				`<path d='M10.896 2H8.75V.75a.75.75 0 0 0-1.5 0V2H5.104a.25.25 0 0 0-.177.427l2.896 2.896a.25.25 0 0 0 .354 0l2.896-2.896A.25.25 0 0 0 10.896 2ZM8.75 15.25a.75.75 0 0 1-1.5 0V14H5.104a.25.25 0 0 1-.177-.427l2.896-2.896a.25.25 0 0 1 .354 0l2.896 2.896a.25.25 0 0 1-.177.427H8.75v1.25Zm-6.5-6.5a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM6 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 6 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM12 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 12 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5Z'/>`,
+				`</svg>`,
+			]),
 		},
 	},
 	cssVarReplacements: [['collapsibleSections', 'cs']],
+	preventUnitlessValues: ['collapsibleSections.closedBorderWidth', 'collapsibleSections.openBorderWidth'],
 })
 
 export function getCollapsibleSectionsBaseStyles({ cssVar }: ResolverContext) {
-	// Icon source: Octicons (MIT licensed)
-	const unfoldSvg = createInlineSvgUrl(
-		'm8.177.677 2.896 2.896a.25.25 0 0 1-.177.427H8.75v1.25a.75.75 0 0 1-1.5 0V4H5.104a.25.25 0 0 1-.177-.427L7.823.677a.25.25 0 0 1 .354 0ZM7.25 10.75a.75.75 0 0 1 1.5 0V12h2.146a.25.25 0 0 1 .177.427l-2.896 2.896a.25.25 0 0 1-.354 0l-2.896-2.896A.25.25 0 0 1 5.104 12H7.25v-1.25Zm-5-2a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM6 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 6 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM12 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 12 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5Z'
-	)
-	const foldSvg = createInlineSvgUrl(
-		'M10.896 2H8.75V.75a.75.75 0 0 0-1.5 0V2H5.104a.25.25 0 0 0-.177.427l2.896 2.896a.25.25 0 0 0 .354 0l2.896-2.896A.25.25 0 0 0 10.896 2ZM8.75 15.25a.75.75 0 0 1-1.5 0V14H5.104a.25.25 0 0 1-.177-.427l2.896-2.896a.25.25 0 0 1 .354 0l2.896 2.896a.25.25 0 0 1-.177.427H8.75v1.25Zm-6.5-6.5a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM6 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 6 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5ZM12 8a.75.75 0 0 1-.75.75h-.5a.75.75 0 0 1 0-1.5h.5A.75.75 0 0 1 12 8Zm2.25.75a.75.75 0 0 0 0-1.5h-.5a.75.75 0 0 0 0 1.5h.5Z'
-	)
 	return `
 		.${collapsibleSectionClass} {
 			position: relative;
@@ -174,8 +194,8 @@ export function getCollapsibleSectionsBaseStyles({ cssVar }: ResolverContext) {
 					}
 				}
 				.expand::after {
-					-webkit-mask-image: ${unfoldSvg};
-					mask-image: ${unfoldSvg};
+					-webkit-mask-image: ${cssVar('collapsibleSections.expandIcon')};
+					mask-image: ${cssVar('collapsibleSections.expandIcon')};
 					/* Ensure that the expand icons of closed sections get printed to avoid gap */
 					-webkit-print-color-adjust: exact;
 					print-color-adjust: exact;
@@ -183,8 +203,8 @@ export function getCollapsibleSectionsBaseStyles({ cssVar }: ResolverContext) {
 				.collapse {
 					display: none;
 					&::after {
-						-webkit-mask-image: ${foldSvg};
-						mask-image: ${foldSvg};
+						-webkit-mask-image: ${cssVar('collapsibleSections.collapseIcon')};
+						mask-image: ${cssVar('collapsibleSections.collapseIcon')};
 					}
 				}
 				.text {
@@ -240,10 +260,4 @@ export function getCollapsibleSectionsBaseStyles({ cssVar }: ResolverContext) {
 			}
 		}
 	`
-}
-
-function createInlineSvgUrl(d: string) {
-	const svg = `<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'><path d='${d}'/></svg>`
-	const encodedSvg = svg.replace(/</g, '%3C').replace(/>/g, '%3E')
-	return `url("data:image/svg+xml,${encodedSvg}")`
 }