+++ packages/@expressive-code/plugin-frames/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @expressive-code/plugin-frames
 
+## 0.41.1
+
+### Patch Changes
+
+- Updated dependencies [a53e749]
+  - @expressive-code/core@0.41.1
+
+## 0.41.0
+
+### Minor Changes
+
+- 380bfcc: Adds new `createInlineSvgUrl` export that creates an inline SVG image data URL from the given contents of an SVG file.
+
+  You can use it to embed SVG images directly into a plugin's styles or HAST, or pass it to an existing `styleOverrides` icon setting.
+
+- 6497f09: Uses the new `preventUnitlessValues` property of `PluginStyleSettings` to make style calculations in the plugins "Collapsible Sections", "Frames" and "Text Markers" more robust.
+- 380bfcc: Adds the following new `styleOverrides` settings:
+
+  - `frames.copyIcon`: Allows overriding the SVG icon used for the copy button. Thank you @louisescher!
+  - `frames.terminalIcon`: Allows overriding the SVG icon used for the terminal window frame. Defaults to three dots in the top left corner.
+
+### Patch Changes
+
+- Updated dependencies [380bfcc]
+- Updated dependencies [0f33477]
+- Updated dependencies [6497f09]
+- Updated dependencies [a826a4a]
+- Updated dependencies [0f33477]
+- Updated dependencies [0f33477]
+  - @expressive-code/core@0.41.0
+
 ## 0.40.2
 
 ### Patch Changes