+++ packages/@expressive-code/plugin-line-numbers/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @expressive-code/plugin-line-numbers
 
+## 0.41.1
+
+### Patch Changes
+
+- Updated dependencies [a53e749]
+  - @expressive-code/core@0.41.1
+
+## 0.41.0
+
+### Patch Changes
+
+- Updated dependencies [380bfcc]
+- Updated dependencies [0f33477]
+- Updated dependencies [6497f09]
+- Updated dependencies [a826a4a]
+- Updated dependencies [0f33477]
+- Updated dependencies [0f33477]
+  - @expressive-code/core@0.41.0
+
 ## 0.40.2
 
 ### Patch Changes